### 2021-06-25 #3400 @whedon
[PRE REVIEW]: Finnish Media Scrapers
**Submitting author:** @jiemakel (<a href="http://orcid.org/0000-0002-8366-8414">Eetu Mäkelä</a>)
**Repository:** <a href="https://github.com/hsci-r/finnish-media-scrapers" target ="_blank">https://github.com/hsci-r/finnish-media-scrapers</a>
**Version:** v1.0.1
