### 2021-06-25 #3400 @kyleniemeyer
Hi @ajstewartlang, could you edit this submission? This looks to align with your interests in open and reproducible research.
