### 2021-06-28 hsci-r/finnish-media-scrapers#7 @fergusq
Support for Python async(io) functions
I have a program where I'd like to use these scrapers, but it uses asynchronous coroutines instead of threads to do multiple simultaneous queries. Unfortunately the scrapers use requests and selenium, which only work in synchronous functions (they will block the program).

I use asynchronous functions because they are much simpler to use in Python that threads and multiprocessing (in my opinion, synchronization, error handling and eg. limiting number of concurrent requests is easier).
