### 2021-06-29 3d66a48
move to using poetry as the build / dependency management environment
