### 2021-06-29 c83bb6f
fix #8 by waiting for the URL to change after submitting credentials
