### 2021-06-29 7785bbd
this sleep should no longer be necessary after adding explicit waiting to the logging in
