### 2021-06-29 hsci-r/finnish-media-scrapers#7 @jiemakel
I'm sort of two minds about this. On the one hand, I want to be mindful of the medias and not make spamming them with concurrent HTTP queries easier (there's already a conservative 1 second average delay between queries in the scripts). On the other hand, I guess this should be left to be up to the user, particularly for programmatic use.

I'll see if I'll do this just as a familiarization with Python coroutines.
