### 2021-06-29 hsci-r/finnish-media-scrapers#7 @fergusq
Just to be clear, in my case I'm making multiple simultaneous queries to different sites, not the same site. Having all requests in the same process makes it easier (imo) to synchronize the coroutines so that each site has its own lock so that only one request is made to the site at time.
