### 2021-06-30 427dfdb
Move to using async coroutines in querying and fetching
