+++ finnish_media_scrapers/scripts/fetch_hs.py
@@ -4,14 +4,14 @@
 # %%
 
 import argparse
+import asyncio
 import csv
 import logging
 import os
 import random
 from time import sleep
 
-from selenium import webdriver
-from selenium.common.exceptions import WebDriverException
+from pyppeteer import launch
 
 from ..fetch import fetch_article_hs, prepare_session_hs
 
@@ -19,7 +19,7 @@ logging.basicConfig(level=logging.INFO)
 # %%
 
 
-def parse_arguments():
+def _parse_arguments():
     parser = argparse.ArgumentParser()
     parser.add_argument(
         '-i', '--input', help="input CSV file containing articles to fetch (from query-hs.py)", required=True)
@@ -32,62 +32,48 @@ def parse_arguments():
     parser.add_argument('-mw', '--max-web-driver-wait',
                         help="maximum time in seconds to wait for the webdriver to render a page before failing (default 30)", default=30, type=int)
     parser.add_argument(
-        '-w', '--web-driver', help="the Selenium web driver to use (out of remote [default],chrome,firefox)", default="remote")
-    parser.add_argument(
         '-d', '--delay', help="number of seconds to wait between consecutive requests (default 1.0)", default=1.0, type=float)
     parser.add_argument('--quiet', default=False,
                         action='store_true', help="Log only errors")
     return parser.parse_args()
 
 
-def main():
-    args = parse_arguments()
+async def _amain():
+    args = _parse_arguments()
     if args.quiet:
         logging.basicConfig(level=logging.ERROR)
+    browser = await launch()
     try:
-        if args.web_driver == 'remote':
-            try:
-                driver = webdriver.Remote(
-                    desired_capabilities=webdriver.DesiredCapabilities.CHROME)
-            except WebDriverException:
-                logging.error(
-                    "Cannot download articles if Selenium is not running on localhost. (e.g. docker run -p 4444:4444 -v /dev/shm:/dev/shm selenium/standalone-chrome:4.0.0-beta-1-20210215)")
-                return
-        elif args.web_driver == 'chrome':
-            driver = webdriver.Chrome()
-        elif args.web_driver == 'firefox':
-            driver = webdriver.Firefox()
-        else:
-            logging.error(
-                "Unknown web driver specified. Must be one of remote/chrome/firefox.")
-            return
-    except WebDriverException:
-        logging.error("Couldn't initialize selected Selenium webdriver.")
-        return
-    try:
-        logging.info("Connecting to Selenium.")
-        logging.info("Logging into HS.")
-        prepare_session_hs(driver, args.username,
-                           args.password, args.max_web_driver_wait)
-        logging.info("Logged in.")
-        os.makedirs(args.output, exist_ok=True)
-        with open(args.input) as input_file:
-            csv_input = csv.DictReader(input_file)
-            for article in csv_input:
-                url = article['url']
-                file = os.path.join(args.output, str(article['id'])+".html")
-                if not os.path.exists(file):
-                    article = fetch_article_hs(
-                        driver, url, args.max_web_driver_wait)
-                    with open(file, "w") as article_file:
-                        article_file.write(
-                            "<!DOCTYPE html><head><meta charset='utf-8'></head>" + article + "</html>")
-                    logging.info("Wrote article %s into %s", url, file)
-                    sleep(random.randrange(args.delay*2))
-                else:
-                    logging.info("Skipping %s as %s already exists.", url, file)
+        session = await browser.newPage()
+        try:
+            logging.info("Logging into HS.")
+            await prepare_session_hs(session, args.username,
+                                     args.password, args.max_web_driver_wait)
+            logging.info("Logged in.")
+            os.makedirs(args.output, exist_ok=True)
+            with open(args.input) as input_file:
+                csv_input = csv.DictReader(input_file)
+                for article in csv_input:
[…]
