+++ finnish_media_scrapers/query.py
@@ -1,11 +1,13 @@
 """Functions related to querying articles from the apis of YLE, Helsingin Sanomat (HS), Ilta-Sanomat (IS) and Iltalehti (IL)
 """
 from datetime import datetime, timedelta
-from typing import Iterator
+from typing import AsyncIterable
 
-import requests
+import attr
+from aiohttp import ClientSession
 
 
+@attr.s
 class Article:
     """An article
 
@@ -15,19 +17,13 @@ class Article:
         title (str): the title or headline of the article
         date_modified (str): the date of last modification for the article
     """
-
-    def __init__(self, id: str, url: str, title: str, date_modified: str):
-        self.id = id
-        self.url = url
-        self.title = title
-        self.date_modified = date_modified
-
-    def __repr__(self) -> str:
-        """Return a string representation of the article
-        """
-        return f"{self.__class__.__name__}((id={self.id},url={self.url},title={self.title},date_modified={self.date_modified})"
+    id: str = attr.ib()
+    url: str = attr.ib()
+    title: str = attr.ib()
+    date_modified: str = attr.ib()
 
 
+@attr.s
 class Result:
     """A result from a single API call
 
@@ -36,25 +32,19 @@ class Result:
         url (str): the URL of the API query
         total (int): the total number of articles for the query. -1 if not available.
     """
-
-    def __init__(self, articles: 'list[Article]', url: str, total: int = -1):
-        self.articles = articles
-        self.url = url
-        self.total = total
-
-    def __repr__(self) -> str:
-        """Return a string representation of the result
-        """
-        return f"{self.__class__.__name__}((url={self.url},articles={self.articles},total={self.total})"
+    articles: 'list[Article]' = attr.ib()
+    url: str = attr.ib()
+    total: int = attr.ib(default=-1)
 
 
 yle_api: str = "https://yle-fi-search.api.yle.fi/v1/search"
 
 
-def query_yle(query: str, language: str, from_date: str, to_date: str, batch_size: int = 10000) -> Iterator[Result]:
+async def query_yle(session: ClientSession, query: str, language: str, from_date: str, to_date: str, batch_size: int = 10000) -> AsyncIterable[Result]:
     """Query the YLE API for articles matching a query
 
     Args:
+        session (ClientSession): the aiohttp session to use
         query (str): the query string to search for
         language (str): language to search (either 'fi' or 'sv')
         from_date (str): date to search from (inclusive, YYYY-MM-DD)
@@ -65,7 +55,7 @@ def query_yle(query: str, language: str, from_date: str, to_date: str, batch_siz
         ValueError: when something goes wrong in the API call
 
     Yields:
-        Iterator[Result]: each Result contains the results from a single API call
+        AsyncIterable[Result]: each Result contains the results from a single API call
     """
     params = {
         'app_id': 'hakuylefi_v2_prod',
@@ -81,42 +71,42 @@ def query_yle(query: str, language: str, from_date: str, to_date: str, batch_siz
         'offset': 0,
         'limit': batch_size
     }
-    response = requests.get(yle_api, params)
-    if response.status_code != 200:
-        raise ValueError(
-            f"Got unexpected response code {response.status_code} for {response.url}.")
-    response_json = response.json()
-    if response_json is None:
-        raise ValueError(f"Got empty response for {response.url}")
-    if response_json['meta']['count'] > 10000:
-        raise ValueError(
-            f"Query results in {response_json['meta']['count']} results. The YLE API refuses to return more than 10000 results, so refusing to continue. You can work around this limitation by doing multiple queries on smaller timespans.")
-    response = requests.get(yle_api, params)
-    while True:
-        if response.status_code != 200:
+    async with session.get(yle_api, params=params) as response:
[…]
