### 2021-06-30 hsci-r/finnish-media-scrapers#7 @jiemakel
This is now implemented in [v1.1.0](https://github.com/hsci-r/finnish-media-scrapers/releases/tag/v1.1.0).
