### 2021-06-30 2fa63c2
fix readthedocs still expecting the project using conda
