+++ .readthedocs.yaml
@@ -12,7 +12,3 @@ sphinx:
 # Optionally build your docs in additional formats such as PDF
 formats:
    - pdf
-
-# Source environment from conda
-conda:
-  environment: environment.yml
