### 2021-06-30 d713d45
add back project metadata that was lost in converting from setuptools to poetry
