### 2021-06-30 707296c
try to fix readthedocs to work with poetry
