+++ .readthedocs.yaml
@@ -12,3 +12,11 @@ sphinx:
 # Optionally build your docs in additional formats such as PDF
 formats:
    - pdf
+
+python:
+  version: 3.7
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
