### 2021-07-16 #3504 @whedon
[REVIEW]: Finnish Media Scrapers
**Submitting author:** @jiemakel (<a href="http://orcid.org/0000-0002-8366-8414">Eetu Mäkelä</a>)
**Repository:** <a href="https://github.com/hsci-r/finnish-media-scrapers" target ="_blank">https://github.com/hsci-r/finnish-media-scrapers</a>
**Version:** v1.1.1

    - Author @jiemakel has made 77 commits, 39,314 ++ lines, 6,319 -- deletions.
    - There are two authors with contributions over at least six months.
    - There are 1,263 lines of code.
    - According to the JOSS draft, this package has been used by two other projects (what are these projects?)

    - Check these:
    - [x] Query YLE/HS/IL/IS APIs for matching articles using the scripts fms-query-{yle|hs|il|is}, which output all matching articles with links into CSVs.
    - [x] Fetch the matching articles using fms-fetch-{hs|open}. These save the articles as HTML files in a specified directory.
    - [x] Extract the plain text from the article HMTL using fms-html-to-text-{yle|svyle|hs|il|is}.
    - [x] Optionally refilter the results using fms-post-filter.

    - There are no performance claims in the JOSS draft. See author's comment [below](https://github.com/openjournals/joss-reviews/issues/3504#issuecomment-909157981) on performance

    - Problems being solved: "responds to research needs by providing easy-to-use tools for scraping Finnish media articles and extracting the article texts from the scraped HTML files."
    - The target audience for this project is those in media research who need access to this data. The python package itself is "for the benefit of more computationally-savvy users" and the "Finish Computational Science community".
    - These is an explicit list of dependencies in [poetry.lock](https://github.com/hsci-r/finnish-media-scrapers/blob/master/poetry.lock)
    - The necessary dependencies are installed with `pip install`
    - [x] Members of Parliament Example:
    - [x] Helsingin Sanomat
    - [x] Yle
    - [x] Iltalehti
    - [x] Iltasanomat
    - [x] fms-post-filter script
    - There are no automated tests due to the reliance on external APIs. Instead, any problems should show through logging. See author's explanation below.
    - I don't think there are clear guidelines aside from the fact that anyone can open a GitHub Issue. EDIT: this has been updated in commit __. In short, all support/contributions can be through GitHub.

    - Yes, there is a section titled 'Statement of Need'
    - The types of problems are: media research
    - The target audience is: researchers, Finnish computational science community
    - There are no named comparable packages, but the Statement of Need addresses the media research field which might use this package.
    - EDIT: a Related work section has been added to the README.

    - Several commits, useful project to academic community studying finnish media, over 1000 LOC  of Python

    - installed via pypi, installed easily
    - `--help` works as expected
    - `fms-query-il` is broken it seems, rest of the tools work (I ran this command: `fms-query-{yle|hs|il|is} --from-date 2021-01-01 -q 'parliament' -l 10 -o results2.csv`). I got a `TypeError` for `-il`.
    - did not check the other three components yet
    - Software's primary bottleneck is not burdening remote servers, therefore, there is a limit on how fast they can be. Regardless, as I understand it, this is a Python script designed more for convenience rather than performance or scalability. Therefore, I see no need for extra efforts in this direction.

    - A clear statement of need is given and target audience (Finnish HSRG) and problem statement (crawl Finnish media text) is given
    - I think it is tradition to also compare one's work with other existing work in the field :-) It would be good to at least mention that no other existing software can possibly be repurposed to perform the same task with the same level of efficiency, as this project directly uses APIs and also extracts structured text, which other softwares will possibly not.
    - There is a related work section in the README but I think a gist of it should still be mentioned in the paper.
    - `pip` is used to install as an end user and `poetry` as a developer, both are pretty standard.
    - Many examples, also an example `members_of_parliament` folder is available.
    - Simple flowchart plus explanatory statements are given.
    - Added a comment about this below (https://github.com/openjournals/joss-reviews/issues/3504#issuecomment-963320985)
    - lead author's email is provided for contact

    - Summary is probably too short and vague. The first step in this project is keyword search but it's not mentioned in the summary. Perhaps a line or two can be added about "structured retrieval" as well. I would begin with something like "Finnish Media Scrapers is a package for retrieving articles in a structured form from Finnish Media websites. It allows flexible configuration in terms of start date/end date and/or keyword search." and then continue to explain what is meant by 1. retrieving articles 2. structured form 3. keywords 4. which media websites.
    - See comment in previous section
