### 2021-08-23 #3504 @Sara-ShiHo
So far, I have a few questions/suggestions for the authors:
 - I believe `.DS_Store` should be removed from the repo and be added to `.gitignore`
 - I have not found any 'performance claims'. I understand that speed might vary due to internet performance, but it might be useful to give a rough idea of how long web scraping takes using this tool. For example, how long will the queries in the example 'Members of parliament' take? Are the run times for the different sources approximately the same?
 - Are there any automated tests?
 - I did not find any 'Community guidelines' regarding third party contributions. I believe that should be added somewhere in order to pass the review.
 - I believe the 'Statement of Need' section in the paper was sufficient. But I was curious whether there are comparable packages for media research which should be mentioned?
