### 2021-08-31 901bbf5
improve documentation in response to JOSS review
