### 2021-08-31 #3504 @jiemakel
In commit https://github.com/hsci-r/finnish-media-scrapers/commit/901bbf522c7451da4928ba4724a69d25ff7605dd, I've tried to answer the questions/suggestions given by @Sara-ShiHo .

More specifically beyond what is now added to the README:
 * on performance, as opposed to trying to be efficient, the scrapers actually by default sleep for one second (a hugely conservative setting) between each request to the media servers in order to be good netizens and not overload their servers. At the same time, the code has actually been refactored to use async io in https://github.com/hsci-r/finnish-media-scrapers/issues/7, which, if wanted, would allow a programmatic user to launch as many simultaneous requests as they dare.
 * on automated tests, there are none. I would like for there to be some, but found them problematic to include and to possibly give a false sense of security. In truth, the internal functioning of the library is quite simple. The complexity goes into interacting with the various media APIs as well as correctly extracting the texts from their varying article layouts. Here, we're anyway at the mercy of the outlets changing their APIs and layouts. Understanding this, my strategy has been to program defensively and to make the crawlers complain loudly through logging whenever they're encountering something unexpected. At the same time, I would have liked to include a set of tests ensuring that current article layouts continue to be properly extracted whenever adding something new. However, that would have required including HTML examples of such articles as part of the repo, and that could have brought IPR concerns, particularly as one of the outlets requires an active subscription to read the articles. So, in the end, I decided the defensive coding fallback should be enough, wherein I can manually rerun the crawler after changes, and it will complain in quite a specific, traceable and debuggable manner if I've broken anything.
