### 2021-09-14 #3504 @Sara-ShiHo
First off, thank you @jiemakel  for your detailed revision. I have a few more suggestions.

* I was hoping that the "Related Work" section would be added to the paper since technically it's in the "Software paper" part of the review checklist (I don't have an opinion on whether it needs to be in the README).
* For the Helsingin Sanomat documenation, add more explicit instructions for username and password such as: "Sign up for username and password at https://www.hs.fi/".
* Have a folder with sample output from one of the media. I think users might like to see that and determine whether it fits their needs before installing.
* I guess the /members_of_parliament folder contains some of the code used for a paper, which makes sense. However, as a standalone example for this repo, I think it takes too long to run. It would be more useful if the example were quicker so that the user could see the output more immediately. Maybe shortening the date range or only querying members of one party will be sufficient to cut down the time.

Finally, I think the paper is well-written and doesn't need change, but I thought the README documentation for HS could be reorganized like this (although feel free to do differently if you disagree).

> This scraper requires a user id and password for Helsingin Sanomat. Sign up for a username and password at https://www.hs.fi/ and add them to the command `fms-fetch-hs`. For example `fms-fetch-hs -i hs-sdp.csv -o hs-sdp -u <username> -p <password>`.
>
> Technically, the `fms-fetch-hs` command uses `pyppeteer` to control a headless Chromium browser to log in and ensure the dynamically rendered content in HS articles is captured. To ensure a compatible Chromium, when first running the tool, `pyppeteer` will download an isolated version of Chromium for itself, causing some ~150MB of network traffic and disk space usage.
>
> After fetching the articles, extract texts `fms-html-to-text` with e.g. `fms-html-to-text-hs -o hs-sdp-output hs-sdp`.
