+++ members_of_parliament/fetch-mp-inflections.py
@@ -3,7 +3,7 @@ import os
 
 import pandas as pd
 
-df = pd.read_csv("mp-expanded.csv")
+df = pd.read_csv("mp-expanded-orpo.csv")
 
 parser = argparse.ArgumentParser()
 parser.add_argument('-u', '--username',
