+++ example-outputs/example-il.txt
@@ -1,0 +1,109 @@
+UutisetPolitiikka
+
+Hallitus valmistelee ulkomaille pelaamisen rajoittamista Veikkauksen tukemiseksi - Saarikko: ”Muotoillaan estoja”
+
+
+
+18.12.2020 klo 15:09
+
+Hallitus on neuvotellut joulun alla arpajaislain muuttamisesta. Keskustan puheenjohtaja Annika Saarikko kertoo Iltalehdelle, mitä hallitus aikoo tehdä. ”Muotoillaan estoja ulkomaille pelaamiseen.”
+
+
+
+
+
+”Haluamme vahvistaa pelaamisen kohdentamista Suomeen niin, että muotoillaan estoja ulkomaille pelaamiseen, jolloin pelaaminen kanavoituisi Veikkauksen kautta”, tiede- ja kulttuuriministeri Annika Saarikko kertoo arpajaislain uudistamisesta. Matti Matikainen
+
+
+
+Tiede- ja kulttuuriministeri, keskustan puheenjohtaja Annika Saarikko kertoo, että hän raaputtaa joka aamu luukun Veikkauksen joulukalenterista. Niin isoa joululahjaa kalenterista ei ole kuitenkaan luvassa, että sillä rahoittaisi koronataantumasta, mainonnan vähentämisestä ja peliautomaattien tunnistautumispakosta seuraavan Veikkauksen tuottovajeen.
+
+Vuonna 2019 Veikkaus tuotti Suomen valtiolle rahaa yli miljardi euroa. Päättyvän vuoden tulos on jäämässä noin 700 miljoonaan euroon, mikä on iso poliittinen päänsärky.
+
+Pelitoiminnan tuotoilla on rahoitettu leijonanosa suomalaisesta liikunnasta, urheilusta, kulttuurista, taiteesta, tieteestä ja järjestöjen toteuttamista sosiaali- ja terveyspalvelujen hankkeista.
+
+Keskustan puheenjohtaja ja hallituksen johtoviisikon jäsen Saarikko kertoo Iltalehdelle, että hallitus on neuvotellut marraskuussa ja joulukuussa aktiivisesti keinoista turvata liikunnan, kulttuurin, taiteen, tieteen ja sosiaali- ja terveysalan järjestöjen rahoitus.
+
+Hallitus valmistelee ulkomaille pelaamisen rajoittamista, jotta Veikkauksen tulot turvataan.
+
+– Arpajaislaki on ollut viime viikot hallituksen neuvotteluissa. Sitä ollaan muokkaamassa. Haluamme vahvistaa pelaamisen kohdentamista Suomeen niin, että muotoillaan estoja ulkomaille pelaamiseen, jolloin pelaaminen kanavoituisi Veikkauksen kautta. Yksi iso ongelma, miksi Veikkauksen tuloutus on pienentynyt, on se, että ihmiset, jotka pelaavat vaikkapa nettipelejä, pelaavat paljon muita kuin Veikkauksen pelejä, Saarikko sanoo.
+
+Saarikko arvioi, että hallituksen on aktiivisesti puolustettava Veikkauksen pelimonopolia. Muussa tapauksessa pelitoiminnan tuotot romahtavat, vaikka monopoli säilyisi paperilla.
+
[…]
