### 2021-09-17 #3504 @UMTti
I've done the following additions to the repository as @Sara-ShiHo suggested:

- Related work section has been added to the paper in this [commit](https://github.com/hsci-r/finnish-media-scrapers/commit/79e3bda4717e74f47985ac817886d46844e10911)
- In the [commit](https://github.com/hsci-r/finnish-media-scrapers/commit/8462e0fc6588bc9a8a7ae7da26587830a0648d91), instructions for signing up to HS has been added
- Example output files have been added as HTML and text files in this [commit](https://github.com/hsci-r/finnish-media-scrapers/commit/0a4e3b307be8930631dfa22a988dc3a0d97b4d2b)
- Members of parliament -example has been changed to download articles only about Petteri Orpo, chairperson of National Coalition Party. Also, the timeframe has been limited. The possibility to download articles about all members of parliament is still existing. ([commit in the example folder](https://github.com/hsci-r/finnish-media-scrapers/commit/68a07a2b724f77f393291d8daf6212d3f6793af4) and [commit in README.md](https://github.com/hsci-r/finnish-media-scrapers/commit/59911f8354b7e40f641a58ea7089185ca1d67f58))
