### 2021-11-08 #3504 @GaurangTandon
Hi, I am happy to review the paper.

For now, my primary concern is of automated tests. I read your response above (https://github.com/openjournals/joss-reviews/issues/3504#issuecomment-909157981). You mentioned that:

> However, that would have required including HTML examples of such articles as part of the repo, and that could have brought IPR concerns

I understand that there are two components to this project: the scraper (which downloads content using API) and the parser (which extracts structured data from the HTML). I believe the authors can _still write valid tests_ for the parser without violating IPR concerns, by simply replacing the original text with some other text (for example, replacing the actual news headline with "This is a news headline" or the body with "This is a news article body"). The parser's primary concern is of the HTML structure whereas the IPR's primary concern is of the news content, and both of these seem to me as mutually satisfiable.

In fact, one can even write a cronjob to "refresh" these HTML files regularly by downloading latest HTML files from the news outlets, and making appropriate replacements to please IPR concerns (would require writing a separate script, though)

Overall, yes, this testing framework would be a bit unusual or perhaps clunky. I don't think not having automated tests are a show-stopper for this project, as there is defensive logging in place as well, but I would appreciate if there were some tests regardless.

I would love to hear what the authors think about this. I still haven't read the project fully so my analysis maybe mistaken, humble apologies if that's the case!

I will follow up with a detailed review soon. Cheers!
