### 2021-11-10 #3504 @GaurangTandon
Hi, in addition to my comment about automated tests above, I have also added some other comments in the main GitHub issue (https://github.com/openjournals/joss-reviews/issues/3504#issue-946176417) about a broken functionality and two textual incoherencies that I think should be addressed.

Now, I'm only left with analysing the functionality of the other three components in this project, I'll analzye them _after_ receiving author's response on the above comments that I have already posted.

Please feel free to reach out to me on email if you want to chat in detail about any of my comments. Thanks!
