### 2021-11-11 c8017d8
make default to-date work for IL query
