### 2021-11-11 #3504 @jiemakel
`fms-query-il` has now been updated to not require a `--to-date` parameter which was what caused the `TypeError`.

On the issue of bypassing IPR concerns by modifying the files for automated tests, I agree. We'll work on adding tests to cover examples of the different HTML formats.
