### 2021-11-25 f068ed5
fix logging in svyle html to text conversion
