### 2021-11-25 6c7bf11
add regression and defensive programming tests
