+++ tests/test_parsing.py
@@ -1,0 +1,139 @@
+"""Regression and defensive programming tests for finnish-media-scrapers
+"""
+
+import pytest
+
+from finnish_media_scrapers.htmltotext import (extract_text_from_hs_html,
+                                               extract_text_from_il_html,
+                                               extract_text_from_is_html,
+                                               extract_text_from_svyle_html,
+                                               extract_text_from_yle_html)
+
+# %%
+
+
+def remove_content(input_file, output_file):
+    """Utility function to remove all content from HTML articles so that no IPR problems remain when using them in tests"""
+    from bs4 import BeautifulSoup
+    with open(input_file) as inf, open(output_file, 'w') as outf:
+        soup = BeautifulSoup(inf, 'lxml')
+        for text_node in soup.find_all(text=True):
+            text_node.replace_with("TEXT")
+        outf.write(str(soup))
+
+# %%
+
+
+def test_hs_parser():
+    """Test parsing of various HS article formats"""
+
+    # common page-main-content variant
+    with open("tests/example_articles/hs/2000006466966-sanitized.html", encoding="utf-8") as htmlf, \
+            open("tests/example_articles/hs/2000006466966-sanitized.txt", encoding="utf-8") as txtf:
+        expected = txtf.read()
+        assert expected == extract_text_from_hs_html(htmlf)
+
+    # paid-content variant
+    with open("tests/example_articles/hs/2000006560604-sanitized.html", encoding="utf-8") as htmlf, \
+            open("tests/example_articles/hs/2000006560604-sanitized.txt", encoding="utf-8") as txtf:
+        expected = txtf.read()
+        assert expected == extract_text_from_hs_html(htmlf)
+
+    # nuxt variant
+    with open("tests/example_articles/hs/2000006585357-sanitized.html", encoding="utf-8") as htmlf, \
+            open("tests/example_articles/hs/2000006585357-sanitized.txt", encoding="utf-8") as txtf:
+        expected = txtf.read()
+        assert expected == extract_text_from_hs_html(htmlf)
+
+
+def test_yle_parser():
+    """Test parsing of various YLE article formats"""
+
+    # common .yle__article variant
+    with open("tests/example_articles/yle/3-11222287-sanitized.html", encoding="utf-8") as htmlf, \
+            open("tests/example_articles/yle/3-11222287-sanitized.txt", encoding="utf-8") as txtf:
+        expected = txtf.read()
+        assert expected == extract_text_from_yle_html(htmlf)
+
+    # .yle__section--article variant
+    with open("tests/example_articles/yle/3-11609882-sanitized.html", encoding="utf-8") as htmlf, \
+            open("tests/example_articles/yle/3-11609882-sanitized.txt", encoding="utf-8") as txtf:
+        expected = txtf.read()
+        assert expected == extract_text_from_yle_html(htmlf)
+
+    # article.content variant
+    with open("tests/example_articles/yle/3-11173365-sanitized.html", encoding="utf-8") as htmlf, \
+            open("tests/example_articles/yle/3-11173365-sanitized.txt", encoding="utf-8") as txtf:
+        expected = txtf.read()
+        assert expected == extract_text_from_yle_html(htmlf)
+
+
+def test_svyle_parser():
+    """Test parsing of various Svenska YLE article formats"""
+
+    # common page-main-content variant
+    with open("tests/example_articles/svyle/7-1456870-sanitized.html", encoding="utf-8") as htmlf, \
+            open("tests/example_articles/svyle/7-1456870-sanitized.txt", encoding="utf-8") as txtf:
+        expected = txtf.read()
+        assert expected == extract_text_from_svyle_html(htmlf)
+
+
[…]
