### 2021-11-25 #3504 @jiemakel
* Regression and defensive programming unit tests added in https://github.com/hsci-r/finnish-media-scrapers/commit/6c7bf11a8524f7e7433354a6b73062f50126af2a
* README and paper updated to better introduce the tool and related work in https://github.com/hsci-r/finnish-media-scrapers/commit/89f5a63969464baf87492178b7a99aa7eafffc0a
