### 2021-12-18 #3504 @GaurangTandon
Hi:

1. The automated testing is clean. Thanks for implementing it!
2. `fms-query-il` works now as per your bugfix.

Overall, I believe all functional and detail checkpoints are complete :heavy_check_mark:

There's one minor issue though: I think the software paper is perhaps too short. Maybe it has not yet been updated and I'm looking at the older version. https://github.com/openjournals/joss-papers/blob/joss.03504/joss.03504/10.21105.joss.03504.pdf
I have highlighted my concerns with it in my review checklist in the parent issue post.

Regardless, the overall GitHub repository itself is complete and comprehensive. Can @ajstewartlang you shed some more light on if these missing points in the software paper are relevant enough or can they be skipped?

Barring this point, my review is complete. Thank you for offering me the opportunity and I wish good luck to the project authors! :-)
