### 2021-12-21 #3504 @UMTti
Hi @GaurangTandon, I added the latest version of the paper [here](https://github.com/hsci-r/finnish-media-scrapers/blob/master/paper_21_12.pdf).
