### 2021-12-21 #3504 @ajstewartlang
@UMTti Can you confirm that the latest version of the paper is this one please? https://github.com/hsci-r/finnish-media-scrapers/blob/master/paper.md
